% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags_mcmc.R
\name{tar_jags_df_draws}
\alias{tar_jags_df_draws}
\title{Modified draws data frame from \code{R2jags} output.}
\usage{
tar_jags_df_draws(x, chains)
}
\arguments{
\item{x}{\code{R2jags} \code{sims.matrix} with posterior draws.}

\item{chains}{Number of MCMC chains.}
}
\description{
Not a user-side function. Do not call directly.
Exported for infrastructure reasons only.
}
\keyword{internal}
