% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags_mcmc.R
\name{tar_jags_df_summary}
\alias{tar_jags_df_summary}
\title{Modified summary data frame from \code{R2jags} output.}
\usage{
tar_jags_df_summary(x)
}
\arguments{
\item{x}{\code{R2jags} summary data frame.}
}
\description{
Not a user-side function. Do not call directly.
Exported for infrastructure reasons only.
}
\keyword{internal}
