% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags_example_data.R
\name{tar_jags_example_data}
\alias{tar_jags_example_data}
\title{Example data for \code{\link[=tar_jags_example_file]{tar_jags_example_file()}}}
\usage{
tar_jags_example_data(n = 10L, true_params = TRUE)
}
\arguments{
\item{n}{Integer of length 1, number of data points.}

\item{true_params}{Logical, whether to include the true values of
the parameters used to simulate the data.}
}
\value{
List, dataset compatible with the model file from
\code{\link[=tar_jags_example_file]{tar_jags_example_file()}}.
}
\description{
An example dataset compatible with the model file
from \code{\link[=tar_jags_example_file]{tar_jags_example_file()}}.
}
\examples{
tar_jags_example_data()
}
