% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags_example_file.R
\name{tar_jags_example_file}
\alias{tar_jags_example_file}
\title{Write an example JAGS model file.}
\usage{
tar_jags_example_file(path = "jagstargets_example.jags")
}
\arguments{
\item{path}{Character of length 1, file path to write the model file.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Overwrites the file at \code{path} with a built-in example
JAGS model file.
}
\examples{
path <- tempfile()
tar_jags_example_file(path = path)
writeLines(readLines(path))
}
