% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags_mcmc.R
\name{tar_jags_summary_tibble}
\alias{tar_jags_summary_tibble}
\title{Clean up the summary data frame from \code{R2jags}.}
\usage{
tar_jags_summary_tibble(x)
}
\description{
Not a user-side function. Do not call directly.
Exported for infrastructure reasons only.
}
\keyword{internal}
