% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_jags.R
\name{tar_jags_run}
\alias{tar_jags_run}
\title{Run a JAGS model and return the whole output object.}
\usage{
tar_jags_run(
  jags_lines,
  parameters.to.save,
  data,
  inits,
  n.cluster,
  n.chains,
  n.iter,
  n.burnin,
  n.thin,
  jags.module,
  RNGname,
  jags.seed,
  stdout,
  stderr,
  progress.bar,
  refresh
)
}
\arguments{
\item{jags_lines}{Character vector of lines from a JAGS model file.}

\item{stdout}{Character of length 1, file path to write the stdout stream
of the model when it runs. Set to \code{NULL} to print to the console.
Set to \code{R.utils::nullfile()} to suppress stdout.
Does not apply to messages, warnings, or errors.}

\item{stderr}{Character of length 1, file path to write the stderr stream
of the model when it runs. Set to \code{NULL} to print to the console.
Set to \code{R.utils::nullfile()} to suppress stderr.
Does not apply to messages, warnings, or errors.}
}
\value{
An \code{R2jags} output object.
}
\description{
Not a user-side function. Do not invoke directly.
}
\keyword{internal}
